[Stef] switch packed type to be a long. pack up to 8 chars.

+++ artio-codecs/src/main/java/uk/co/real_logic/artio/dictionary/SessionConstants.java
@@ -41,13 +41,13 @@ public final class SessionConstants
     public static final int NEW_PASSWORD = 925;
 
 
-    public static final int LOGON_MESSAGE_TYPE = 65;
-    public static final int HEARTBEAT_MESSAGE_TYPE = 48;
-    public static final int TEST_REQUEST_MESSAGE_TYPE = 49;
-    public static final int RESEND_REQUEST_MESSAGE_TYPE = 50;
-    public static final int REJECT_MESSAGE_TYPE = 51;
-    public static final int SEQUENCE_RESET_MESSAGE_TYPE = 52;
-    public static final int LOGOUT_MESSAGE_TYPE = 53;
+    public static final long LOGON_MESSAGE_TYPE = 65;
+    public static final long HEARTBEAT_MESSAGE_TYPE = 48;
+    public static final long TEST_REQUEST_MESSAGE_TYPE = 49;
+    public static final long RESEND_REQUEST_MESSAGE_TYPE = 50;
+    public static final long REJECT_MESSAGE_TYPE = 51;
+    public static final long SEQUENCE_RESET_MESSAGE_TYPE = 52;
+    public static final long LOGOUT_MESSAGE_TYPE = 53;
 
     public static final int USER_REQUEST_MESSAGE_TYPE = 17730;
