+++ artio-samples/src/main/java/uk/co/real_logic/artio/builder/OrderSingleEncoder.java
@@ -109,9 +109,9 @@ public class OrderSingleEncoder implements Encoder
         return this;
     }
 
-    public int messageType()
+    public long messageType()
     {
-        return 0;
+        return 0L;
     }
 
     public SessionHeaderEncoder header()
